+++ src/main/java/com/rcloud/server/sealtalk/dao/GroupFavsMapper.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface GroupFavsMapper extends Mapper<GroupFavs> {
 
-    List<GroupFavs> queryGroupFavsWithGroupByUserId(@Param("userId") Integer userId, @Param("limit") Integer limit, @Param("offset") Integer offset);
+    List<GroupFavs> queryGroupFavsWithGroupByUserId(@Param("userId") Long userId, @Param("limit") Integer limit, @Param("offset") Integer offset);
 
-    Integer queryCountGroupFavsWithGroupByUserId(@Param("userId") Integer userId);
+    Integer queryCountGroupFavsWithGroupByUserId(@Param("userId") Long userId);
 }
