+++ src/main/java/com/rcloud/server/sealtalk/dao/GroupMembersMapper.java
@@ -8,11 +8,11 @@ import java.util.List;
 
 public interface GroupMembersMapper extends Mapper<GroupMembers> {
 
-    List<GroupMembers> queryGroupMembersWithGroupByMemberId(@Param("memberId") Integer memberId);
+    List<GroupMembers> queryGroupMembersWithGroupByMemberId(@Param("memberId") Long memberId);
 
     List<GroupMembers> queryGroupMembersWithUsersByGroupId(@Param("groupId") String groupId,@Param("isDeleted") Integer isDeleted);
 
-    GroupMembers queryGroupMembersWithGroupByGroupIdAndMemberId(@Param("groupId") String groupId, @Param("memberId") Integer memberId);
+    GroupMembers queryGroupMembersWithGroupByGroupIdAndMemberId(@Param("groupId") String groupId, @Param("memberId") Long memberId);
 
     List<GroupMembers> selectGroupMembersWithUsersByGroupIdsAndVersion(@Param("groupIdList") List<String> groupIdList, @Param("version") Long version);
